move settings out into context to reduce number of arguments
