+++ nhpy/run_detailed_results.py
@@ -37,8 +37,6 @@ import gc
 import os
 import sys
 import time
-from ast import Assert
-from contextlib import suppress
 from logging import INFO
 from pathlib import Path
 from typing import List
@@ -66,6 +64,7 @@ from nhpy.utils import (
     _load_dotenv_file,
     configure_logging,
     get_logger,
+    initialise_connections_and_params,
 )
 
 # %%
@@ -94,64 +93,6 @@ def get_memory_usage():
 
 # Load environment variables
 _load_dotenv_file(interpolate=False)
-
-
-def _initialise_connections_and_params(
-    results_path: str,
-    account_url: str,
-    results_container: str,
-    data_container: str,
-) -> ProcessContext:
-    """
-    Initialize connections and load parameters from the aggregated results.
-
-    Args:
-        results_path: Path to the aggregated model results
-        account_url: Azure Storage account URL
-        results_container: Azure Storage container for results
-        data_container: Azure Storage container for data
-
-    Returns:
-        dict containing all necessary objects and parameters for processing
-
-    Raises:
-        FileNotFoundError: If results folder or data version not found
-    """
-    # Connections and params
-    results_connection = az.connect_to_container(account_url, results_container)
-    data_connection = az.connect_to_container(account_url, data_container)
-    params = az.load_agg_params(results_connection, results_path)
-
-    # Get info from the results file
-    scenario_name = params["scenario"]
-    trust = params["dataset"]
-    model_version = params["app_version"]
-    baseline_year = int(params["start_year"])
-    run_id = params["create_datetime"]
-
-    # Patch model version for loading the data. Results folder name truncated,
-    # e.g. v3.0 does not show the patch version. But data stores in format v3.0.1
-    model_version_data = az.find_latest_version(data_connection, params["app_version"])
-    logger.info(f"Using data: {model_version_data}")
-    if model_version_data == "N/A":
-        raise FileNotFoundError("Results folder not found")
-
-    # Add principal to the "vanilla" model results
-    actual_results_df = az.load_agg_results(results_connection, results_path)
-    actual_results_df = process_results.convert_results_format(actual_results_df)
-
-    return {
-        "results_connection": results_connection,
-        "data_connection": data_connection,
-        "params": params,
-        "scenario_name": scenario_name,
-        "trust": trust,
-        "model_version": model_version,
-        "model_version_data": model_version_data,
-        "baseline_year": baseline_year,
-        "run_id": run_id,
-        "actual_results_df": actual_results_df,
-    }
 
 
 def _check_results_exist(output_dir: str, scenario_name: str, activity_type: str) -> bool:
@@ -619,51 +560,9 @@ def _process_aae_results(
     )
 
 
-def suppress_small_counts(
-    df: pd.DataFrame,
-    suppress_cols: List[str],
-    count_col: str = "baseline",
-    threshold: int = 5,
-) -> pd.DataFrame:
-    """Suppression of small counts in detailed results. Filters dataframe to only rows with
-    small numbers, then groups together values in specified columns and re-aggregates.
-
-    Args:
-        df (pd.DataFrame): Processed detailed results
-        suppress_cols (List[str]): List of columns to use in suppressing small numbers,
-        in order of preference.
-        count_col (str, optional): The name of the column with the values to be suppressed
-        and aggregated.
-        Defaults to "value".
-        threshold (int, optional): Maximum count allowed in the count_col
-
-    Returns:
-        pd.DataFrame: Processed detailed results with small counts aggregated together
-    """
-    logger.info("Beginning suppression...")
-    full_index_cols = df.index.names.copy()
-    for col in suppress_cols:
-        df = df.reset_index()
-        keep = df[df[count_col] >= threshold].copy()
-        small = df[df[count_col] < threshold].copy()
-        # avoid suppression if we don't need to
-        if small.empty:
-            df = df.groupby(full_index_cols).sum()
-            break
-        small[col] = "grouped"
-        # reaggregate
-        grouped = small.groupby(full_index_cols, dropna=False).sum().reset_index()
-        # bind suppressed rows with the ones that didn't need suppression
-        df = pd.concat([keep, grouped], ignore_index=True).groupby(full_index_cols).sum()
-    return df.sort_index().sort_index(axis=1)
-
-
 def run_detailed_results(
-    results_path: str,
+    context: ProcessContext,
     output_dir: str | None = None,
-    account_url: str | None = None,
-    results_container: str | None = None,
-    data_container: str | None = None,
     agg_type: str = "standard",
 ) -> dict[str, str]:
     """
@@ -674,11 +573,9 @@ def run_detailed_results(
     exist for any activity type, those processing steps are skipped.
 
     Args:
-        results_path: Path to existing aggregated results
+        context: ProcessContext with metadata for specific scenario to run
         output_dir: Directory to save output files (default: 'nhpy/data')
-        account_url: Azure Storage account URL (default: from environment)
-        results_container: Azure Storage container for results (default: from environment)
-        data_container: Azure Storage container for data (default: from environment)
+        agg_type: Type of aggregation. These are set in config
 
     Returns:
         dict[str, str]: dictionary containing paths to output files
@@ -692,34 +589,11 @@ def run_detailed_results(
     # Start the total timing
     total_start_time = time.perf_counter()
 
-    # Load environment variables if not provided
-    account_url = account_url or os.getenv("AZ_STORAGE_EP", "")
-    results_container = results_container or os.getenv("AZ_STORAGE_RESULTS", "")
-    data_container = data_container or os.getenv("AZ_STORAGE_DATA", "")
-
-    if not all([account_url, results_container, data_container]):
-        missing = []
-        if not account_url:
-            missing.append("AZ_STORAGE_EP")
-        if not results_container:
-            missing.append("AZ_STORAGE_RESULTS")
-        if not data_container:
-            missing.append("AZ_STORAGE_DATA")
-        raise EnvironmentVariableError(
-            missing_vars=missing,
-            message=f"Missing environment variables: {', '.join(missing)}",
-        )
-
     # Set up output directory
     if output_dir is None:
         output_dir = "nhpy/data"
 
     Path(output_dir).mkdir(parents=True, exist_ok=True)
-
-    # Initialise connections and load parameters
-    context = _initialise_connections_and_params(
-        results_path, account_url, results_container, data_container
-    )
 
     scenario_name = context["scenario_name"]
 
@@ -816,6 +690,16 @@ def main() -> int:
 
     args = parser.parse_args()
 
+    # Initialise context and connections
+    account_url = args.account_url or os.getenv("AZ_STORAGE_EP", "")
+    results_container = args.results_container or os.getenv("AZ_STORAGE_RESULTS", "")
+    data_container = args.data_container or os.getenv("AZ_STORAGE_DATA", "")
+
+    # Initialise context just to get scenario name
+    context = initialise_connections_and_params(
+        args.results_path, account_url, results_container, data_container
+    )
+
     try:
         # First, check if specific result type was requested
         if args.ip or args.op or args.ae:
@@ -825,17 +709,6 @@ def main() -> int:
             # We need to set up the output directory
             Path(output_dir).mkdir(parents=True, exist_ok=True)
 
-            # Initialize connections to get scenario name
-            account_url = args.account_url or os.getenv("AZ_STORAGE_EP", "")
-            results_container = args.results_container or os.getenv(
-                "AZ_STORAGE_RESULTS", ""
-            )
-            data_container = args.data_container or os.getenv("AZ_STORAGE_DATA", "")
-
-            # Initialise context just to get scenario name
-            context = _initialise_connections_and_params(
-                args.results_path, account_url, results_container, data_container
-            )
             scenario_name = context["scenario_name"]
 
             # Check for existing files for the requested type
@@ -852,11 +725,8 @@ def main() -> int:
         # If we're here, either no specific type was requested or files don't exist yet
         # Run detailed results generation
         result_files = run_detailed_results(
-            results_path=args.results_path,
+            context,
             output_dir=args.output_dir,
-            account_url=args.account_url,
-            results_container=args.results_container,
-            data_container=args.data_container,
             agg_type=args.agg_type,
         )
